[go/users] Fix GetFollowers and GetFollowings APIs logic.

+++ go/users/service/users.go
@@ -116,6 +116,11 @@ func (s *usersSvc) UpdateBio(ctx context.Context, p *users.UpdateBioPayload) err
 func (s *usersSvc) Follow(ctx context.Context, p *users.FollowPayload) error {
 	followed_user_id, _ := uuid.Parse(p.FollowedUserID)
 	following_user_id, _ := uuid.Parse(p.FollowingUserID)
+	if following_user_id == followed_user_id {
+		err := users.MakeBadRequest(errors.New("user tries to follow self"))
+		s.logger.Printf("users.Follow: failed (%s)", err)
+		return err
+	}
 	err := s.followshipsRepo.CreateFollowship(ctx, followed_user_id, following_user_id)
 	if err != nil {
 		s.logger.Printf("users.Follow: failed (%s)", err)
@@ -129,6 +134,11 @@ func (s *usersSvc) Follow(ctx context.Context, p *users.FollowPayload) error {
 func (s *usersSvc) Unfollow(ctx context.Context, p *users.UnfollowPayload) error {
 	followed_user_id, _ := uuid.Parse(p.FollowedUserID)
 	following_user_id, _ := uuid.Parse(p.FollowingUserID)
+	if following_user_id == followed_user_id {
+		err := users.MakeBadRequest(errors.New("user tries to unfollow self"))
+		s.logger.Printf("users.Unfollow: failed (%s)", err)
+		return err
+	}
 	err := s.followshipsRepo.DeleteFollowship(ctx, followed_user_id, following_user_id)
 	if err != nil {
 		s.logger.Printf("users.Unfollow: failed (%s)", err)
